+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - 40: Feedback forms and reports
+- Accessibility improvements
+- Add multiple severity levels to sitewide alert
 - Fix: StreamField typo causing RecursionError in Wagtail 2.13
 - Use the newly released django-gov-notify library for GOV.UK Notify integration